% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualizing_MAGMA_functions.R
\name{Balance_MAGMA}
\alias{Balance_MAGMA}
\title{Balance estimation}
\usage{
Balance_MAGMA(Data, group, covariates, step = "step")
}
\arguments{
\item{Data}{A data frame containing at least the \emph{grouping} variable, the
\emph{step} variable from the main MAGMA-function (or other mathcing algorithms),
and all \emph{covariates} of interest.}

\item{group}{A character specifying the name of
your grouping variable in data. Note that MAGMA can only match your data for
a maximum of 4 groups. For matching over two groups (e.g., 2x2 Design) is
possible by specifying group as a character vector with a length of two. In
this case each or the 2 grouping variables can only have two levels.}

\item{covariates}{A character vector listing the names of all binary and
metric covariates of interest.}

\item{step}{A character naming the step variable of the matching. Per
default it specifies the name as \emph{step}, which corresponds the resulting
name of the main MAGMA function.}
}
\value{
A list of length four containing all balance criteria and all
pairwise effects with respect to group sample size.
}
\description{
This function computes all four Balance criteria of MAGMA, namely
\emph{Pillai's Trace}, \emph{d-ratio}, \emph{mean g}, and \emph{adjusted d-ratio}. It can
consider binary and metric variables for balance estimation. Balance
Estimation is conducted over a varying sample size. See Details for more
information.
}
\details{
This function computes all four Balance criteria of MAGMA, namely Pillai's
Trace, d-ratio, mean g, and adjusted d-ratio. This is an iterative process
including more cases with each iteration according to the step variable.
Thus, starting with cases having a small within match distance, larger
distances are included with increasing iterations. As minimum the function
specifies a n >= 20 per group. This does not imply, that balance criteria
with such a small sample size can be estimated consistently. For Pillai's
Trace a higher minimum sample size can be possible. It depends on the number
of covariates to ensure a positive model identification.
Missing data for Pillai's Trace are excluded listwise, whiel for the other
balance criteria pairwise exclusion is applied.
}
\examples{
\dontrun {
#Defining the names of the metric and binary covariates
covariates_vector <- c("GPA_school", "IQ_score", "Motivation", "parents_academic", "sex")

#Matching data set for giftedness support with exact matching for enrichment.
MAGMA_sim_data_gifted_exact <- MAGMA_exact(Data = MAGMA_sim_data,
                                           group = "gifted_support",
                                           dist = "ps_gifted",
                                           exact = "enrichment",
                                           cores = 2)


#Estimating Balance
Balance_gifted_exact <- Balance_MAGMA(Data = MAGMA_sim_data_gifted_exact,
                                      group = "gifted_support",
                                      covariates = covariates_vector,
                                      step = "step") #step created during matching
str(Balance_gifted_exact)

#Computing 2x2 Matching for giftedness support and enrichment equivalent to
#a four group matching
MAGMA_sim_data_gift_enrich <- MAGMA(Data = MAGMA_sim_data,
                                   group = c("gifted_support", "enrichment"),
                                   dist = "ps_2x2",
                                   cores = 2)


#Estimating Balance. Covariates same as above
Balance_2x2 <- Balance_MAGMA(Dta = MAGMA_sim_data_gift_enrich,
                             group = c("gifted_support", "enrichment"),
                             covariates = covariates_vector,
                             step = "step") #step created during matching
str(Balance_2x2)

}

}
\author{
Julian Urban
}
