% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Density_overlap.R
\name{Density_overlap}
\alias{Density_overlap}
\title{Density_overlap}
\usage{
Density_overlap(
  Data,
  variable,
  group,
  variable_name = NULL,
  group_labels = NULL,
  group_name = NULL,
  step_num = NULL,
  step_var = NULL
)
}
\arguments{
\item{Data}{A data frame that contains the desired variable for density
plotting as well as the specified grouping variable.}

\item{variable}{A character naming the variable for that the density should
be plotted.}

\item{group}{A character specifying the groups for which the density should
be plotted. Can be an independent group comparison (e.g., comparing matched
groups) or the comparison of pre and post matched samples.}

\item{variable_name}{A character specifying the name that should appear
in the plot for variable.}

\item{group_labels}{A character vector specifying the labels for the groups
that should appear in the legend of the plot.}

\item{group_name}{A character specifying the name of the grouping variable
tha should appear in the title of the legend.}

\item{step_num}{An integer specifying the number of cases that should be included
per group in this post matching comparison. Is based on the step variable of
MAGMA.}

\item{step_var}{A Character specifying the name of the step variable.}
}
\value{
A Plot showing the kernel density for a specified variable separated
for specified groups and the quantification of this overlap.
}
\description{
This function shows and quantifies the kernel density overlap of a variable
for two or more groups.
}
\details{
This function enables the comparison of the density of variables for two
or more groups. It plots the kernel density separately for each group and
quantifies the amount of overlap.
}
\examples{
\dontrun {
#Density overlap in Propensity scores for gifted before matching
Density_overlap(Data = MAGMA_sim_data,
variable = "ps_gifted", 
group = "gifted_support",
step = NULL,
variable_name = "Propensity Score",
group_labels = c("No Support", "Support"),
group_name = "Gifted Support")

 #Density overlap in Propensity scores for gifted after matching with 250 cases per group.
Density_overlap(Data = MAGMA_sim_data_gifted,
variable = "ps_gifted", 
group = "gifted_support",
step_num = 250,
step_var = "step",
variable_name = "Propensity Score",
group_labels = c("No Support", "Support"),
group_name = "Gifted Support")

#Density overlap in GPA for teacher ability rating before matching
Density_overlap(Data = MAGMA_sim_data,
variable = "GPA_school", 
group = "teacher_ability_rating",
variable_name = "School Achievement",
group_labels = c("Low", "Medium", "High"),
group_name = "Rating") 

}

}
\references{
\insertRef{Pastore M, Loro PAD, Mingione M, Calcagni' A (2022). \emph{overlapping: Estimation of Overlapping in Empirical Distributions}. R package version
2.1, \url{https://CRAN.R-project.org/package=overlapping}.}
}
\author{
Julian Urban
}
