% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualizing_MAGMA_functions.R
\name{initial_unbalance}
\alias{initial_unbalance}
\title{Initial Unbalance estimation}
\usage{
initial_unbalance(Data, group, covariates)
}
\arguments{
\item{Data}{A data frame containing at least the \emph{grouping} variable and all
\emph{covariates} of interest.}

\item{group}{A character specifying the name of
your grouping variable in data. Note that MAGMA can only match your data for
a maximum of 4 groups. For matching over two groups (e.g., 2x2 Design) is
possible by specifying group as a character vector with a length of two. In
this case each or the 2 grouping variables can only have two levels.}

\item{covariates}{A character vector listing the names of all binary and
metric covariates of interest.}
}
\value{
A numeric vector of length 4 containg the respective balance
criteria for the unmatched sample.
}
\description{
This function computes all four Balance criteria of MAGMA, namely
\emph{Pillai's Trace}, \emph{d-ratiO}, \emph{mean g}, and \emph{adjusted d-ratio} for the
unmatched data set. This enables comparision of inital and after matching
unbalance.
}
\details{
This function computes all four Balance criteria of MAGMA, namely Pillai's
Trace, d-ratio, mean g, and adjusted d-ratio for the overall samples. This
leads to different sample sizes per group. Missing data for Pillai's Trace
are excluded listwise, whiel for the other balance criteria pairwise
exclusion is applied.
}
\examples{
\dontrun {
#Defining covariates for balance estimation
covariates_vector <- c("GPA_school", "IQ_score", "Motivation", "parents_academic", "sex")

#Computing initial unbalance for giftedness support
unbalance_gifted <- initial_unbalance(Data = MAGMA_sim_data,
                                      group = "gifted_support",
                                      covariates = covariates_vector)
unbalance_gifted

#Computing initial unbalnce for teacher rated ability
unbalance_tar <- initial_unbalance(Data = MAGMA_sim_data,
                                  group = "teacher_ability_rating",
                                  covariates = covariates_vector)
unbalance_tar

#' #Computing initial unbalnce for teacher rated ability
unbalance_2x2 <- initial_unbalance(Data = MAGMA_sim_data,
                                  group = c("gifted_support", "enrichment"),
                                  covariates = covariates_vector)
unbalance_2x2
}

}
\author{
Julian Urban
}
