% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_g_functions.R
\name{inner_variance}
\alias{inner_variance}
\title{inner variance}
\usage{
inner_variance(d_square, size_per_group)
}
\arguments{
\item{d_square}{The squared Cohen's d.}

\item{size_per_group}{Specifies the group size.}
}
\value{
A numeric specifying the variance of an effect.
}
\description{
inner function of mean_g.
}
\details{
This function computed the variance of a Cohen's d.
}
\author{
Julian Urban
}
