% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAGMA_function.R
\name{MAGMA}
\alias{MAGMA}
\title{MAGMA}
\usage{
MAGMA(Data, group, dist, cores = 1)
}
\arguments{
\item{Data}{A data frame or tibble containing at least your grouping and
distance variable. Data needs to be specified in your environment.}

\item{group}{A character specifying the name of
your grouping variable in data. Note that MAGMA can only match your data for
a maximum of 4 groups. For matching over two groups (e.g., 2x2 Design) is
possible by specifying group as a character vector with a length of two. In
this case each or the 2 grouping variables can only have two levels.}

\item{dist}{A character specifying the name of your distance
variable in data.}

\item{cores}{An integer defining the number of cores used for
parallel computation.}
}
\value{
Your input data frame augmented with matching
relevant variables, namely \emph{weight}, \emph{step}, \emph{distance}, and \emph{ID}. In case
of missing values on the distance or group variable, MAGMA excludes them for
the matching process. The returned data set does not contain those excluded
cases. For more information, see Details.
}
\description{
This function conducts many group matching for 2 to 4 groups. It augments
the original data set by relevant MAGMA variables. For details, see below.
}
\details{
This function conducts nearest neighbor many group matching. It is
applicable for 2 to 4 groups or a 2x2 Design. As output, this function
augment your original data by the variables \emph{weight}, \emph{step}, \emph{distance},
and \emph{ID}. Weight indicates whether a case was matched. Step specifies the
iteration in which a case was matched. It also shows which cases were matched
together. Distance indicates the mean difference within a match. Since
matches with a lower distance are matched in an earlier iteration, step and
distance are strongly correlated.
This function has some CPU and RAM load. In most four group applications and
three group applications with large sample size, RAM may be not sufficient.
Therefore MAGMA switches to random quasi-systematic matching. If this is the
case MAGMA informs you. Results does not change, but Balance might be
slightly affected.
}
\examples{
{
\dontrun{
#Running this code will take a while
#Computing two-group Matching for giftedness support
MAGMA_sim_data_gifted<- MAGMA(Data = MAGMA_sim_data,
                                group = "gifted_support",
                                dist = "ps_gifted",
                                cores = 2)
head(MAGMA_sim_data_gifted)

#Conducting three-group matching for teacher ability rating. Cores per
#default = 1
MAGMA_sim_data_tar <- MAGMA(Data = MAGMA_sim_data,
                            group = "teacher_ability_rating",
                            dist = "ps_tar")
head(MAGMA_sim_data_tar)

#Computing 2x2 Matching for giftedness support and enrichment equivalent to
#a four group matching
MAGMA_sim_data_gift_enrich <- MAGMA(Data = MAGMA_sim_data,
                                   group = c("gifted_support", "enrichment"),
                                   dist = "ps_2x2",
                                   cores = 2)
head(MAGMA_sim_data_gift_enrich)
}
}

}
\author{
Julian Urban
}
