% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Desc_function.R
\name{MAGMA_desc}
\alias{MAGMA_desc}
\title{MAGMA_desc}
\usage{
MAGMA_desc(
  Data,
  covariates,
  group,
  step_num = NULL,
  step_var = NULL,
  filename = NULL
)
}
\arguments{
\item{Data}{A data frame that contains the desired variable for density
plotting as well as the specified grouping variable.}

\item{covariates}{A character vector naming the variable names of the
continuous varibales for that the descriptive statistics should be computed.}

\item{group}{A character (vector) specifying the groups for which
differentiated statistics should be computed.}

\item{step_num}{Optional argument. If no value is specified, pre matching
statistics are computed. For post matching statistics the exact needs to be
defined via the stp arguments. step_num is andinteger specifying the number
of cases that should be included per group in this post matching comparison.
Is based on the step variable of MAGMA.}

\item{step_var}{Optional argument. If no value is specified, pre matching
statistics are computed. For post matching statistics the exact needs to be
defined via the stp arguments.A Character specifying the name of the step
variable in the data set.}
}
\value{
A table of descriptvive statistics and pairwise effects for pre- or
postmatching samples.
}
\description{
This function provides pre- and post matching descriptive statistics and
effects.
}
\details{
This function enables the compuatation of descriptive statistics of
continuous variables for the overall sample and specifed groups. Additional,
pairwise effects (Cohen's d) is computed.
}
\examples{
\dontrun {
#Defining covariates
covariates_gifted <- c("GPA_school", "IQ_score", "Motivation", "parents_academic", "sex")

#Estimating pre matching descriptive statistics and pairwise effects
MAGMA_desc(Data = MAGMA_sim_data,
           covariates = covariates_gifted)
           group =  "gifted_support")
           
#Matching the data for gifted support
MAGMA_sim_data_gifted <- MAGMA(Data = MAGMA_sim_data,
                               group = "gifted_support",
                               dist = "ps_gifted",
                               cores = 2)
                               
#Estimating post matching descriptive statistics and pairwise effects
MAGMA_desc(Data = MAGMA_sim_data_gifted,
           covariates = covariates_gifted)
           group =  "gifted_support"
           step_num = 100,
           step_var = "step")

}

}
\author{
Julian Urban
}
