% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAGMA_exact_function.R
\name{MAGMA_exact}
\alias{MAGMA_exact}
\title{MAGMA exact}
\usage{
MAGMA_exact(Data, group, dist, exact, cores = 1)
}
\arguments{
\item{Data}{A data frame or tibble containing at least your grouping and
distance variable. Data needs to be specified in your environment.}

\item{group}{A character specifying the name of
your grouping variable in data. Note that MAGMA can only match your data for
a maximum of 4 groups. For matching over two groups (e.g., 2x2 Design) is
possible by specifying group as a character vector with a length of two. In
this case each or the 2 grouping variables can only have two levels.}

\item{dist}{A character specifying the name of your distance
variable in data.}

\item{exact}{A character specifying the name of the exact variable.
Only cases with the same value on this variable can be matched.}

\item{cores}{An integer defining the number of cores used for
parallel computation.}
}
\value{
Your input data frame of valid cases augmented with matching
relevant variables, namely \emph{weight}, \emph{step}, \emph{distance}, and \emph{ID}. In case
of missing values on the distance or group variable, MAGMA excludes them for
th matching process. The returned data set does not contain those excluded
cases. For more information, see Details.
}
\description{
This function conducts exact many group matching for 2 to 4 groups. Exact
means, that only cases with the same value on the exact variable can be
matched. It #' augments the original data set by relevant MAGMA variables.
For details, see below.
}
\details{
This function conducts nearest neighbor exact many group matching. It is
applicable for 2 to 4 groups or a 2x2 Design. As output, this function
augment your original data by the variables \emph{weight}, \emph{step}, \emph{distance},
and \emph{ID}. Weight indicates whether a case was matched. Step specifies the
iteration in which a case was matched. It also shows which cases were
matched together. Distance indicates the mean difference within a match.
Since matches with a lower distance are matched in an earlier iteration,
step and distance are strongly correlated.
Exact matching means that only cases with the same value on the exact
variable can be matched. As example, only person of the same gender, the
same school, or the same organization are possible matches.
}
\examples{
#Running this code will take a while
#Computing two-group Matching for giftedness support with exact for enrichment
MAGMA_sim_data_gifted_exact <- MAGMA_exact(Data = MAGMA_sim_data,
                                           group = "gifted_support",
                                           dist = "ps_gifted",
                                           exact = "enrichment",
                                           cores = 2)
head(MAGMA_sim_data_gifted_exact)

#Conducting three-group matching for teacher ability rating exact for
#sex. Cores per default = 1
MAGMA_sim_data_tar_exact<- MAGMA_exact(Data = MAGMA_sim_data,
                                       group = "teacher_ability_rating",
                                       dist = "ps_tar",
                                       exact = "sex")
head(MAGMA_sim_data_tar_exact)

#Computing 2x2 Matching for giftedness support and enrichment equivalent to
#a four group matching with exact MAGMA_exact for teacher rated ability
MAGMA_sim_data_gift_enrich_exact <- MAGMA_exact(Data = MAGMA_sim_data,
                                                group = c("gifted_support", "enrichment"),
                                                dist = "ps_2x2",
                                                exact = "teacher_ability_rating",
                                                cores = 2)
head(MAGMA_sim_data_gift_enrich_exact)

}
\author{
Julian Urban
}
