% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Package_documentation.R
\docType{package}
\name{MAGMA_package}
\alias{MAGMA}
\alias{MAGMA-package}
\title{MAGMA: A package for MAny Group MAtching}
\description{
This package enables nearest neighbor matching for two to four groups. It is
also possible to match for a 2x2 Design. Additional to standard matching, we
included exact matching. Besides the matching functions, this
package include a framwork to estimate pre- and post matching balance. This
Framwork includes four balance criteria, namely \emph{Pillai's Trace}, \emph{d-ratio},
\emph{mean g}, and \emph{adjusted d-ratio}. It is possible display these balance
criteria in an APA table. Lastly, one function creates plots to show the
balance trend for different matching scenarios.
}
\author{
Julian Urban \email{julian.urban@gesis.org}
}
