% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Inner_MAGMA_functions.R
\name{match_iterative}
\alias{match_iterative}
\title{Inner MAGMA-matching process function}
\usage{
match_iterative(distance_input, output_list, rep_elements)
}
\arguments{
\item{distance_input}{distance matrix to extract lowest distance}

\item{output_list}{output where MAGMA results get stored}

\item{rep_elements}{sample sizes per group}
}
\value{
A matched sample including the variables step, weight & distance
}
\description{
matches Cases iteratively during MAGMA
}
\details{
This function conducts the matching process, by extracting the match with
the lowest distance
}
