% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_g_functions.R
\name{multiply}
\alias{multiply}
\title{multiply}
\usage{
multiply(x, y)
}
\arguments{
\item{x}{A numeric}

\item{y}{A numeric}
}
\value{
A product.
}
\description{
inner function of mean_g.
}
\details{
This function multiplies two numerics.
}
\author{
Julian Urban
}
