% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_g_functions.R
\name{multiply_variance}
\alias{multiply_variance}
\title{multiply for variance}
\usage{
multiply_variance(x, y)
}
\arguments{
\item{x}{A numeric}

\item{y}{A numeric}
}
\value{
A product.
}
\description{
inner function of mean_g.
}
\details{
This function multiplies a numeric with another squared numeric.
}
\author{
Julian Urban
}
