% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pillai_functions.R
\name{Pillai_DV}
\alias{Pillai_DV}
\title{Pillai DV creation}
\usage{
Pillai_DV(data, input)
}
\arguments{
\item{data}{Specified by Pillai iterativ equlas its data input.}

\item{input}{Specified by Pillai iterativ naming the covariates.}
}
\value{
A matrix of the covariates
}
\description{
DV matrix creator.
}
\details{
This function creates the DV matrix for Pillai iterativ.
}
\author{
Julian Urban
}
