% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualizing_MAGMA_functions.R
\name{Plot_MAGMA}
\alias{Plot_MAGMA}
\title{Balance plots}
\usage{
Plot_MAGMA(
  Balance,
  criterion = c("Pillai", "d_ratio", "mean_g", "Adj_d_ratio")
)
}
\arguments{
\item{Balance}{A Balance MAGMA object. Compare the function Balance_MAGMA.}

\item{criterion}{A character vector specifying for which balance criteria
a plot should be created. Default is all criteria.}
}
\value{
R Plots showing the balance trend over sample size.
}
\description{
Plots for Balance with respect to sample size.
}
\details{
This function creates R-Plots using ggplot to show the belance trend over
sample size.
}
\examples{
\dontrun {
#This function bases on a MAGMA function as well as Balance_MAGMA.
#To run examples, copy them into your console or script.
#Defining the names of the metric and binary covariates
covariates_vector <- c("GPA_school", "IQ_score", "Motivation", "parents_academic", "sex")

#Matching data set for giftedness support with exact matching for enrichment.
MAGMA_sim_data_gifted_exact <- MAGMA_exact(Data = MAGMA_sim_data,
                                           group = "gifted_support",
                                           dist = "ps_gifted",
                                           exact = "enrichment",
                                           cores = 2)


#Estimating Balance
Balance_gifted_exact <- Balance_MAGMA(data = MAGMA_sim_data_gifted_exact,
                                      group = "gifted_support",
                                      covariates = covariates_vector,
                                      step = "step") #step created during matching

Plot_MAGMA(Balance = Balance_gifted_exact) #Using default to plot all criteria

#Computing 2x2 Matching for giftedness support and enrichment equivalent to
#a four group matching
MAGMA_sim_data_gift_enrich <- MAGMA(Data = MAGMA_sim_data,
                                   group = c("gifted_support", "enrichment"),
                                   dist = "ps_2x2",
                                   cores = 2)


#Estimating Balance. Covariates same as above
Balance_2x2 <- Balance_MAGMA(data = MAGMA_sim_data_gift_enrich,
                             group = c("gifted_support", "enrichment"),
                             covariates = covariates_vector,
                             step = "step") #step created during matching

Plot_MAGMA(Balance = Balance_2x2,
           criterion = c("d_ration", "Adj_d_ratio"))
}
           

}
\author{
Julian Urban
}
