% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualizing_MAGMA_functions.R
\name{Table_MAGMA}
\alias{Table_MAGMA}
\title{Balance table}
\usage{
Table_MAGMA(Balance, filename = NULL)
}
\arguments{
\item{Balance}{A result of Balance_MAGMA. Compare the function
\code{\link{Balance_MAGMA}}.}

\item{filename}{Optional argument.  A character specifying the filename that
the resulting Word document with the Table should have.}
}
\value{
A 4x5 APA Table showing the four balance
criteria and the respective sample size per group for four scenarios. In
each of these scenario one balance criteria has its optimal value. It can
print a Word Document with this table, too.
}
\description{
This function prints an APA Table of the Balance criteria. It displays the
balance criteria for four different sample sizes per group. In each scenario
one balance criteria has its optimal value. Thus, the table is a 4x5 table
showing the four balance criteria and the respective sample size per group
for the four scenarios.
}
\details{
This function creates an APA Table including the optimal models for each
balance criterion, the other criteria for this sample size per group as well
as the sample size itself. With an optional argument you can save a the APA
table in Word.
}
\examples{
\dontrun{
#This function bases on a MAGMA function as well as Balance_MAGMA.
#To run examples, copy them into your console or script.
#Defining the names of the metric and binary covariates
covariates_vector <- c("GPA_school", "IQ_score", "Motivation", "parents_academic", "sex")

#Matching data set for giftedness support with exact matching for enrichment.
MAGMA_sim_data_gifted_exact <- MAGMA_exact(Data = MAGMA_sim_data,
                                           group = "gifted_support",
                                           dist = "ps_gifted",
                                           exact = "enrichment",
                                           cores = 2)


#Estimating Balance
Balance_gifted_exact <- Balance_MAGMA(Data = MAGMA_sim_data_gifted_exact,
                                      group = "gifted_support",
                                      covariates = covariates_vector,
                                      step = "step") #step created during matching

Table_MAGMA(Balance_gifted_exact, "Balance_gifted_exact.docx")

#Computing 2x2 Matching for giftedness support and enrichment equivalent to
#a four group matching
MAGMA_sim_data_gift_enrich <- MAGMA(Data = MAGMA_sim_data,
                                   group = c("gifted_support", "enrichment"),
                                   dist = "ps_2x2",
                                   cores = 2)


#Estimating Balance. Covariates same as above
Balance_2x2 <- Balance_MAGMA(Data = MAGMA_sim_data_gift_enrich,
                             group = c("gifted_support", "enrichment"),
                             covariates = covariates_vector,
                             step = "step") #step created during matching

Table_MAGMA(Balance_2x2, "Balance_2x2.docx")
}

}
\author{
Julian Urban
}
